// utils/src/enums.ts
export enum Logger_MessageDirection {
  MessageLeaving,
  MessageEntering
}

export enum Logger_LogLevels {
  None,
  Server,
  ServerMessages,
  Support,
  Gui,
  Development,
  Basic
}